# demo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "email": "yinjie_luo@intsig.net"
        }
    },
    "host": "http://192.168.6.50:10006",
    "basePath": "/edapi/bankbills/",
    "paths": {
        "/edapi/bankbills/analyse/jyds/counterparty/class": {
            "get": {
                "description": "交易对手-对手方分类",
                "tags": [
                    "交易对手"
                ],
                "parameters": [
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "对手方分类展示 0:汇总 1:收入 2:支出 默认展示汇总页",
                        "name": "disp_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.JydsClassResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/jyds/counterparty/overview": {
            "get": {
                "description": "交易对手-对手方概览",
                "tags": [
                    "交易对手"
                ],
                "parameters": [
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "对手方概览展示 0:汇总 1:收入 2:支出 默认展示汇总页",
                        "name": "disp_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.JydsOverviewResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/analyse/jyds/monitor_object": {
            "get": {
                "description": "交易对手-重点监测对象",
                "tags": [
                    "交易对手"
                ],
                "parameters": [
                    {
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "对手方分类展示 0:汇总 1:收入 2:支出 默认展示汇总页",
                        "name": "disp_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接 不超过30个",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 2022-07-24",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "description": "默认5条/页",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "第一页开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目id",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间 2022-07-23",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.JydsMonitorObjectResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/edapi/bankbills/doc/delete": {
            "get": {
                "description": "文件管理-删除指定文档",
                "tags": [
                    "文件管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "待删除文档id",
                        "name": "doc_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.Result1"
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/doc/detail": {
            "get": {
                "description": "文件管理-文档详情",
                "tags": [
                    "文件管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "待查看文档id",
                        "name": "doc_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.DocDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/doc/query": {
            "get": {
                "description": "文件管理-项目文件列表",
                "tags": [
                    "文件管理"
                ],
                "parameters": [
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "文件上传时间排序搜索项 0:默认倒序 1:正序",
                        "name": "date_desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前项目下筛选的文档 多个以英文逗号','拼接",
                        "name": "doc_ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间 秒级时间戳",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "表头校验搜索项 0:不通过 1:通过",
                        "name": "header_check_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字搜索 文件名称或者备注模糊匹配",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "description": "默认5条/页",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            -1,
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "ocr识别状态 -1:识别失败 0:识别中 1:识别成功",
                        "name": "ocr_status",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "第一页开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间 秒级时间戳",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.Result1"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rest.DocListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/edapi/bankbills/doc/update": {
            "get": {
                "description": "文件管理-修改某个文档",
                "tags": [
                    "文件管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "待编辑文档",
                        "name": "doc_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文档备注",
                        "name": "note",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文档名",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.Result1"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rest.ClassTradeInfo": {
            "type": "object",
            "properties": {
                "in_amount": {
                    "description": "收入金额",
                    "type": "integer"
                },
                "name": {
                    "description": "对手方名称",
                    "type": "string"
                },
                "out_amount": {
                    "description": "支出金额",
                    "type": "integer"
                }
            }
        },
...
        "rest.DocDetailResp": {
            "type": "object",
            "properties": {
                "doc_info": {
                    "description": "文档详情信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.DocInfoI"
                        }
                    ]
                },
                "last_doc_id": {
                    "description": "按时间顺序的上个文档id",
                    "type": "string"
                },
                "next_doc_id": {
                    "description": "按时间顺序的下个文档id",
                    "type": "string"
                }
            }
        },
        "rest.DocInfoI": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "项目创建时间",
                    "type": "integer"
                },
                "doc_id": {
                    "description": "文档id",
                    "type": "string"
                },
                "page_list": {
                    "description": "文档下页信息 一个文档可能是多页的",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.DocPageListI"
                    }
                },
                "title": {
                    "description": "文档名",
                    "type": "string"
                },
                "upload_time": {
                    "description": "文档上传时间",
                    "type": "integer"
                }
            }
        },
        "rest.DocListI": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "项目创建时间",
                    "type": "integer"
                },
                "doc_id": {
                    "description": "文档id",
                    "type": "integer"
                },
                "first_page_id": {
                    "description": "若该文档包含多页 这是第一页id",
                    "type": "string"
                },
                "header_check_status": {
                    "description": "表头校验结果",
                    "type": "integer"
                },
                "modify_time": {
                    "description": "文档变更时间",
                    "type": "integer"
                },
                "note": {
                    "description": "项目备注",
                    "type": "string"
                },
                "ocr_status": {
                    "description": "ocr识别结果",
                    "type": "integer"
                },
                "title": {
                    "description": "文档名",
                    "type": "string"
                },
                "upload_time": {
                    "description": "文档上传时间",
                    "type": "integer"
                }
            }
        },
        "rest.DocListResp": {
            "type": "object",
            "properties": {
                "file_list": {
                    "description": "文档详情",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.DocListI"
                    }
                },
                "file_num": {
                    "description": "文档数",
                    "type": "integer"
                }
            }
        },
        "rest.DocPageListI": {
            "type": "object",
            "properties": {
                "file_name": {
                    "description": "页名",
                    "type": "string"
                },
                "modify_time": {
                    "description": "页变更时间",
                    "type": "integer"
                },
                "note": {
                    "description": "页备注",
                    "type": "string"
                },
                "page_id": {
                    "description": "页id",
                    "type": "string"
                },
                "rotate": {
                    "type": "integer"
                },
                "title": {
                    "description": "页标题",
                    "type": "string"
                }
            }
        },
...
        "rest.JydsClassResp": {
            "type": "object",
            "properties": {
                "class": {
                    "description": "分类",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.JydsClassSummary"
                        }
                    ]
                }
            }
        },
        "rest.JydsClassSummary": {
            "type": "object",
            "properties": {
                "institutional_clients": {
                    "description": "机构客户数量",
                    "type": "integer"
                },
                "natural_person": {
                    "description": "自然人数量",
                    "type": "integer"
                },
                "other": {
                    "description": "其他数量",
                    "type": "integer"
                },
                "own": {
                    "description": "自己数量",
                    "type": "integer"
                },
                "trade_info": {
                    "description": "对手方交易明细 top8",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ClassTradeInfo"
                    }
                }
            }
        },
        "rest.JydsMonitorObjectResp": {
            "type": "object",
            "properties": {
                "monitor": {
                    "description": "监测",
                    "allOf": [
                        {
                            "$ref": "#/definitions/rest.JydsMonitorObjectSummary"
                        }
                    ]
                }
            }
        },
        "rest.JydsMonitorObjectSummary": {
            "type": "object",
            "properties": {
                "financial": {
                    "description": "金融服务",
                    "type": "integer"
                },
                "investment": {
                    "description": "投资",
                    "type": "integer"
                },
                "loan": {
                    "description": "贷款",
                    "type": "integer"
                },
                "other": {
                    "description": "其他",
                    "type": "integer"
                },
                "real_estate": {
                    "description": "房地产",
                    "type": "integer"
                },
                "total_num": {
                    "description": "所有流水的条数",
                    "type": "integer"
                },
                "trade_info": {
                    "description": "所有流水详情",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.MonitorTradeInfo"
                    }
                }
            }
        },
        "rest.JydsOverviewResp": {
            "type": "object",
            "properties": {
                "trade_info": {
                    "description": "交易信息 top15+其他 数组最大长度16",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.JydsOverviewTradeInfo"
                    }
                }
            }
        },
        "rest.JydsOverviewTradeInfo": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "交易账号",
                    "type": "string"
                },
                "amount": {
                    "description": "交易金额",
                    "type": "integer"
                },
                "end_time": {
                    "description": "交易时间段结束 2023-02-18",
                    "type": "string"
                },
                "id": {
                    "description": "序号",
                    "type": "integer"
                },
                "name": {
                    "description": "交易户名",
                    "type": "string"
                },
                "start_time": {
                    "description": "交易时间段开始 2022-08-09",
                    "type": "string"
                },
                "trade_number": {
                    "description": "交易笔数",
                    "type": "integer"
                }
            }
        },
...
        "rest.MonitorTradeInfo": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "交易账号",
                    "type": "string"
                },
                "amount": {
                    "description": "交易金额",
                    "type": "integer"
                },
                "end_time": {
                    "description": "交易时间段结束 2023-02-18",
                    "type": "string"
                },
                "name": {
                    "description": "交易户名",
                    "type": "string"
                },
                "signal": {
                    "description": "预警信号",
                    "type": "string"
                },
                "start_time": {
                    "description": "交易时间段开始 2022-08-09",
                    "type": "string"
                },
                "trade_number": {
                    "description": "交易笔数",
                    "type": "integer"
                }
            }
        },
...
        "rest.Result1": {
            "type": "object",
            "properties": {
                "data": {},
                "err": {},
                "ret": {
                    "type": "integer"
                }
            }
        },
...
    }
}